% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateRegex.R
\name{AggregateRegex}
\alias{AggregateRegex}
\title{Inputs a dataframe, and outputs a data frame}
\usage{
AggregateRegex(df)
}
\arguments{
\item{df}{The input data frame}
}
\value{
df
}
\description{
Inputs a dataframe, and outputs a data frame
}
\examples{
AggregateRegex(df)
}
\keyword{Java}
\keyword{load}
