% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyNLPModels.R
\name{ApplyNLPModels}
\alias{ApplyNLPModels}
\title{ApplyNLPModels}
\usage{
ApplyNLPModels(finding.list, text.dfm, regex.df)
}
\arguments{
\item{df}{Input data frame}

\item{site}{1= , 2= , 3=, 4=; defaults to 2}
}
\value{
A folder of files as in all.predsput_filepath that contains dataChunk_1.csv, ...
}
\description{
This function segments the imagereporttext column of a data frame into columns containing text
history, exam, comparison, technique, finding, impression
}
\keyword{chunk}
\keyword{file}
