% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucSummary.R
\name{aucSummary}
\alias{aucSummary}
\title{custom AUC loss function instead of default misclassification error.}
\usage{
aucSummary(data, lev = NULL, model = NULL)
}
\value{
A list
}
\description{
custom AUC loss function instead of default misclassification error.
}
\examples{
aucSummary(data)
}
\keyword{ML,}
\keyword{machine-learning}
