% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcMetrics.R
\name{CalcMetrics}
\alias{CalcMetrics}
\title{calculate auc using trapezoid method.}
\usage{
CalcMetrics(test, truth, N, control, case)
}
\value{
a list
}
\description{
calculate auc using trapezoid method.
}
\examples{
CalcMetrics(test, truth, N, control, case)
}
\keyword{ML,}
\keyword{machine-learning}
