% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTrapezoidalAUC.R
\name{CalcTrapezoidalAUC}
\alias{CalcTrapezoidalAUC}
\title{calculate auc using trapezoid method.}
\usage{
CalcTrapezoidalAUC(sens, spec)
}
\value{
value that represents the AUC
}
\description{
calculate auc using trapezoid method.
}
\examples{
CalcTrapezoidalAUC(sens, spec)
}
\keyword{ML,}
\keyword{machine-learning}
