% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Loads a Java executable file}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{name of library}

\item{pkgname}{name of package}
}
\description{
This function loads a java executable file contained in inst/java folder
}
\examples{
.onLoad()
}
\keyword{Java}
\keyword{load}
