% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegexNegex.R
\name{GetRegexNegex}
\alias{GetRegexNegex}
\title{Inputs a dataframe, and outputs a data frame}
\usage{
GetRegexNegex(df)
}
\arguments{
\item{df}{The input data frame}
}
\description{
Inputs a dataframe, and outputs a data frame
}
\examples{
GetRegexNegex("Jason")
}
\keyword{Java}
\keyword{load}
