% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearningNLP.R
\name{MachineLearningNLP}
\alias{MachineLearningNLP}
\title{Apply machine-learning NLP for the given list of findings}
\usage{
MachineLearningNLP(finding.list, text.dfm, regex.df.wide,
  grouping_var = "imageid")
}
\arguments{
\item{finding.list}{The vector of string of findings}

\item{grouping_var}{Unique identifier of report}
}
\value{
A dataframe containing machine-learing predicted probabilities (between 0 and 1)
and predicted classes (0 or 1) for each finding in finding.list
}
\description{
This function applies the feature weights stored in the ml_model_coef.xlsx file (one sheet = one model)
To a the input feature matrix text.dfm preprocessed using CreateTextFeatures
}
\examples{
MachineLearningNLP(finding.list, text.dfm, regex.df.wide, grouping_var = "imageid")
}
\keyword{ML,}
\keyword{machine-learning}
