% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlMethodSetup.R
\name{mlMethodSetup}
\alias{mlMethodSetup}
\title{This function uses the LireNLPSystem's rule-based functions to classify each report for the finding of interest}
\usage{
mlMethodSetup(df, imageid = "imageid", seed = 187)
}
\arguments{
\item{df}{input dataframe}

\item{imageid}{column indicating the id of each report, default set to "imageid"}

\item{seed}{indicates the seed, default 187}
}
\value{
returns a list of the MLMETHOD, the metric to select probability cutoff, MyControl which is needed for caret, and
        the train and test IDs
}
\description{
This function uses the LireNLPSystem's rule-based functions to classify each report for the finding of interest
}
\examples{
mlMethodSetup(df = outcome.df, imageid = "imageid", seed = 187)
}
\keyword{mlMethodSetup}
