% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleBasedNLP.R
\name{RuleBasedNLP}
\alias{RuleBasedNLP}
\title{Inputs a dataframe, and outputs a data frame}
\usage{
RuleBasedNLP(regex.df.java)
}
\arguments{
\item{regex.df.java}{The input data frame, which is an output from Java JAR file}
}
\value{
List of
regex.df.wide: A data frame with nrow = unique imageids of regex.df.java, having regex (keyword present)
and negex (keyword present but negated)
columns for each of the findings
rules.nlp.df: A data frame with nrow = unique imageids of regex.df.java, having columns for the
report level rules based NLP predictions for each finding
}
\description{
Inputs a dataframe, and outputs a data frame
}
\examples{
RuleBasedNLP(df)

}
\keyword{Java}
\keyword{load}
