% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMLMethod.R
\name{runMLMethod}
\alias{runMLMethod}
\title{This function elastic net logistic regression for a given finding and feature matrix}
\usage{
runMLMethod(finding, featureMatrix, outcome, trainID, testID, metric,
  mlmethod, myControl, outpath)
}
\arguments{
\item{finding}{string indicating finding of interest}

\item{featureMatrix}{feature matrix}

\item{outcome}{dataframe indicating the labels for each report}

\item{trainID}{vector indicating the imageid's for the training data}

\item{testID}{vector indicating the imageid's for the test data}

\item{metric}{string indicating which metric to maximize (eg. "auc" or "fscore")}

\item{mlmethod}{string indicating the ml method (eg. "glmnet")}

\item{myControl}{caret myControl hyperparameters}

\item{outpath}{string indicating outpath to write results to to}
}
\value{
a list of 3 dataframes: metrics (performance metrics for train and test), predictions (predictions on train and test), model's features and coefficients
}
\description{
This function elastic net logistic regression for a given finding and feature matrix
}
\examples{
runMLMethod(finding = "fracture",
            featureMatrix = featureMatrix,
            outcome = outcome.df,
            trainID = c(1,2,3),
            testID = c(4,5,6),
            metric = "auc",
            mlmethod = "glmnet",
            outpath = "Results/")
}
\keyword{runMLMethod}
