% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayHello.R
\name{sayHello}
\alias{sayHello}
\title{Say hello to a user}
\usage{
sayHello(username)
}
\value{
A string greeting the user as called from the Java method
}
\description{
This is a test function to test the rJava interface
whenever changes to Java code is made
This code will be removed from final package release
}
\examples{
sayHello("Jason")
}
\keyword{Java}
\keyword{load}
