% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionSegmentation.R
\name{SectionSegmentation}
\alias{SectionSegmentation}
\title{Section Segmentation of Radiology Text Reports}
\usage{
SectionSegmentation(df, site = NULL, hst_rgx = NULL, exam_rgx = NULL,
  comparison_rgx = NULL, technique_rgx = NULL, body_rgx = NULL,
  imp_rgx = NULL, datetime_rgx = NULL,
  liretext_rgx = "(Among\\\\s*people\\\\s*(over|under|between)\\\\s*the\\\\s*age\\\\s*of.*$)|(Some\\\\s*findings\\\\s*are\\\\s*so\\\\s*common.*$)|(The\\\\s*following\\\\s*findings\\\\s*are\\\\s*so\\\\s*common\\\\s*in\\\\s*normal.*$)")
}
\arguments{
\item{df}{Input data frame}

\item{site}{1= , 2= , 3=, 4=; defaults to 2}

\item{hst_rgx}{Defaults to regex for Kaiser}

\item{exam_rgx}{Defaults to regex for Kaiser}

\item{comparison_rgx}{Defaults to regex for Kaiser}

\item{technique_rgx}{Defaults to regex for Kaiser}

\item{body_rgx}{Defaults to regex for Kaiser}

\item{imp_rgx}{Defaults to regex for Kaiser}

\item{liretext_rgx}{Defaults to regular expressions for LIRE macro; don't change unless there's a good reason to.}
}
\value{
The input data frame with additional columns exam, comparison, technique, body, impression.
The body column is always non-empty: the entire imagereporttext is returned in body if separation is not possible.
}
\description{
This function segments the imagereporttext column of a data frame into columns containing text
history, exam, comparison, technique, finding, impression
}
\examples{
SectionSegmentation(input.df, site = 2)
}
\keyword{section}
\keyword{segmentation}
