% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateRegex.R
\name{AggregateRegex}
\alias{AggregateRegex}
\title{Inputs a dataframe, and outputs a data frame}
\usage{
AggregateRegex(regex.df.java)
}
\arguments{
\item{regex.df.java}{The input data frame, which is an output from Java JAR file}
}
\value{
df
}
\description{
Inputs a dataframe, and outputs a data frame
}
\examples{
AggregateRegex(df)

}
\keyword{Java}
\keyword{load}
