% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegexNegex.R
\name{GetRegexNegex}
\alias{GetRegexNegex}
\title{Inputs a dataframe, and outputs a data frame}
\usage{
GetRegexNegex(annotated.df, imageid = "patientID", siteID = "siteID",
  imageTypeID = "imageTypeID", bodyText = "body",
  impressionText = "impression")
}
\arguments{
\item{annotated.df}{The input data frame}

\item{imageid}{String of column that indexes report images; defaults to patientID}

\item{siteID}{String of column that indexes study site; defaults to siteID}

\item{imageTypeID}{String of column that indexes imaging modality; defaults to imageTypeID}

\item{bodyText}{String of column that indexes body of report text}

\item{bodyText}{String of column that indexes impression sectino of report text}
}
\value{
A "long" data frame with the columns
 Finding
 imageid: ID that indexes reports
 siteID: study site
 imageTypeID: imaging modality
 Sentence: Exact string of the sentence
 Section of sentence: Whether the sentence was in body or impression
 regex: 1 if there is a keyword in the sentence, otherwise 0
 negex: 1 if the keyword is negated, otherwise 0
 **NOTE**: regex cannot be 1 if regex = 0
 keyword: The keyword that triggered regex = 1
}
\description{
Inputs a dataframe, and outputs a data frame
}
\examples{
GetRegexNegex(df)
}
\keyword{Java}
\keyword{load}
